refactor: share the mutation-fns spine via sync-core
The offline outbox's mutation-fns — which replay each optimistic transaction
against the server through tRPC — were duplicated between web and mobile:
identical wrapTrpcError, the createTask auto-suffix loop, and the per-entity
mutate bodies (tasks/messages/resources/properties/teams/seen), differing only
by import paths. This is the last of the parallel client layers ARCHITECTURE.md
§10 / §12.5 flags.

Adds makeCoreMutationFns(trpc) + wrapTrpcError to @buildinlime/sync-core. The
injected client is typed via a CoreTrpc interface built from
inferRouterInputs<contract AppRouter>; because both apps' routers .input() from
the same contracts, each app's real tRPC client is assignable to it, so the
shared payloads stay type-checked (property/seen casts now target domain enums,
not `as any`).

web mutation-fns 335 -> 7 lines (the surface is exactly the core spine); mobile
443 -> 128 (composes the core with its outbox-only entities: projects, build
units, channels). CoreMutationFns is a `type` so it keeps an implicit index
signature and stays assignable to the executor's Record<string, MutationFn>.

web typecheck 221, mobile 13 (both unchanged, zero new errors). All unit tests pass.
